requirements.txt
fastapi[standard]
pydantic
pydantic_settings
uvicorn
sqlalchemy
alembic
asyncpg
greenlet
black
pyjwt
passlib[argon2]
redis
fastapi-cache2
